share data across frames
can now share data across frames. in particular, when the user wants to generate a custom password and have the newly generated password fill inside the entry fields for first and second password entry. There needs to be some edits to handling invalid user data and only deleting the fields that were invalid so the user doesn't have to reenter all fields.