CRUD navigation semi polished
the CRUD Ui frames work for the user to navigate to and from. There are some small bugs that I haven't been able to resolve at this point with the navigation buttons 'back' and the frame instance not being completely destroyed. Otherwise this is a good spot to start on user input validation for add and edit accounts.